# Description

First part of Work Product 12, adding the functionality to start `RunReaders` at an offset.

The new `rawPageFindKey` function has the same performance characteristics as `rawPageLookup` but facilitates starting the `RunReaders` at an give page offset from the returned index. `RunReader.new` now takes an optional offset as an argument.