@mheinzel 2024-08-22
There is a [build failure for GHC versions before 9.6](https://github.com/IntersectMBO/lsm-tree/actions/runs/10493900676/job/29068935700?pr=346#step:13:151) which seems related to `UnboxedSums`.

Also, there is a [docspec-related error](https://github.com/IntersectMBO/lsm-tree/actions/runs/10493900676/job/29068937398?pr=346#step:17:103), seems to happen on all Ubuntu builds.